+++ packages/orq-rc/src/models/operations/createdataset.ts
@@ -138,7 +138,7 @@ export const CreateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-06-25T06:34:33.478Z",
+    "2026-06-25T07:42:51.475Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {